KNF, and 64-bit safety.
SCCS-vsn: sys/vm/vm_page.c 8.4
SCCS-vsn: sys/vm/vm_kern.c 8.4
SCCS-vsn: sys/vm/vm_fault.c 8.5
SCCS-vsn: sys/vm/vm_object.c 8.6
SCCS-vsn: sys/vm/vm_page.h 8.3
SCCS-vsn: sys/vm/vm_pageout.c 8.6
SCCS-vsn: sys/vm/vm_unix.c 8.2
SCCS-vsn: sys/vm/vm_pageout.h 8.3
SCCS-vsn: sys/vm/vm_param.h 8.2
SCCS-vsn: sys/vm/vm_extern.h 8.3

+++ usr/src/sys/vm/vm_kern.c
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)vm_kern.c	8.3 (Berkeley) %G%
+ *	@(#)vm_kern.c	8.4 (Berkeley) %G%
  *
  *
  * Copyright (c) 1987, 1990 Carnegie-Mellon University.
@@ -54,8 +54,8 @@
  *	Allocate pageable memory to the kernel's address map.
  *	map must be "kernel_map" below.
  */
-
-vm_offset_t kmem_alloc_pageable(map, size)
+vm_offset_t
+kmem_alloc_pageable(map, size)
 	vm_map_t		map;
 	register vm_size_t	size;
 {
@@ -83,7 +83,8 @@ vm_offset_t kmem_alloc_pageable(map, size)
  *	Allocate wired-down memory in the kernel's address map
  *	or a submap.
  */
-vm_offset_t kmem_alloc(map, size)
+vm_offset_t
+kmem_alloc(map, size)
 	register vm_map_t	map;
 	register vm_size_t	size;
 {
@@ -175,7 +176,8 @@ vm_offset_t kmem_alloc(map, size)
  *	with kmem_alloc, and return the physical pages
  *	associated with that region.
  */
-void kmem_free(map, addr, size)
+void
+kmem_free(map, addr, size)
 	vm_map_t		map;
 	register vm_offset_t	addr;
 	vm_size_t		size;
@@ -196,7 +198,8 @@ void kmem_free(map, addr, size)
  *	min, max	Returned endpoints of map
  *	pageable	Can the region be paged
  */
-vm_map_t kmem_suballoc(parent, min, max, size, pageable)
+vm_map_t
+kmem_suballoc(parent, min, max, size, pageable)
 	register vm_map_t	parent;
 	vm_offset_t		*min, *max;
 	register vm_size_t	size;
@@ -355,7 +358,8 @@ kmem_malloc(map, size, canwait)
  *	has no room, the caller sleeps waiting for more memory in the submap.
  *
  */
-vm_offset_t kmem_alloc_wait(map, size)
+vm_offset_t
+kmem_alloc_wait(map, size)
 	vm_map_t	map;
 	vm_size_t	size;
 {
@@ -376,7 +380,7 @@ vm_offset_t kmem_alloc_wait(map, size)
 			vm_map_unlock(map);
 			return (0);
 		}
-		assert_wait((int)map, TRUE);
+		assert_wait(map, TRUE);
 		vm_map_unlock(map);
 		thread_block();
 	}
@@ -391,14 +395,15 @@ vm_offset_t kmem_alloc_wait(map, size)
  *	Returns memory to a submap of the kernel, and wakes up any threads
  *	waiting for memory in that map.
  */
-void	kmem_free_wakeup(map, addr, size)
+void
+kmem_free_wakeup(map, addr, size)
 	vm_map_t	map;
 	vm_offset_t	addr;
 	vm_size_t	size;
 {
 	vm_map_lock(map);
 	(void) vm_map_delete(map, trunc_page(addr), round_page(addr + size));
-	thread_wakeup((int)map);
+	thread_wakeup(map);
 	vm_map_unlock(map);
 }
 
@@ -408,7 +413,8 @@ void	kmem_free_wakeup(map, addr, size)
  * map the range between VM_MIN_KERNEL_ADDRESS and `start' as allocated, and
  * the range between `start' and `end' as free.
  */
-void kmem_init(start, end)
+void
+kmem_init(start, end)
 	vm_offset_t start, end;
 {
 	register vm_map_t m;
